feat: Add findByName method to UniversityRepository
The code changes include adding a new method `findByName` to the `UniversityRepository` interface and implementing it in the `PrismaUniversityRepository` class. This method allows finding a university by its name. The purpose of this change is to provide a way to search for universities based on their name.

Note: The recent user commits and repository commits are not relevant for generating the commit message.