chore: Update npm dependencies
Update the npm dependencies in the server/package.json and client/package-lock.json files. This includes updating the versions of "@prisma/client" and "prisma" to "^5.18.0" in the server/package.json file. Additionally, the client/package-lock.json file includes updates to various packages such as "date-fns" and "fs.realpath". These updates ensure that the project is using the latest stable versions of the dependencies, improving stability and security.