refactor: Add 'Teacher' font to login screen
Add the 'Teacher' font to the login screen by modifying the styleLogin.ts file. This change includes adding the 'fontFamily' property with the value 'Teacher' to the 'Uni' style. The 'Teacher' font file is located at '../../../assets/fonts/Teachers Students.otf'. This update improves the visual appearance and consistency of the login screen.